Split all drives into separate files
