--- src/main/kotlin/teksturepako/pakkuDesktop/app/actions/ActionRunner.kt
@@ -1,11 +1,0 @@
-/*
- * Copyright (c) Juraj Hrivnák. All Rights Reserved unless otherwise explicitly stated.
- */
-
-package teksturepako.pakkuDesktop.app.actions
-
-/**
- * ActionRunner is no longer used in the fractal architecture.
- * Actions are launched via [teksturepako.pakkuDesktop.app.ui.driver.LocalLaunchAction]
- * provided by the actionDriver. This file is kept as an empty placeholder.
- */
